tmp/ref.sub.html

<!DOCTYPE html>
<html>
	<head>
		<meta charset="utf-8">
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<script src="./resources/utils.sub.js"></script>
	</head>
	<body>
<script>
'use strict';
const URL_COMPLETE = new URL('/tmp/resources/test-frame.html', location).href;
const URL_DOMAIN = location.origin + '/';

const createIframe = async (t, policy) => {
	const iframe = document.createElement('iframe');
	const pipe = policy ? `?pipe=header(Referrer-Policy,${policy})` : '';
	iframe.src = `${URL_DOMAIN}tmp/resources/test-frame.html${pipe}`;

	await new Promise((resolve, reject) => {
		addEventListener('message', function handler (event) {
			if (event.source !== iframe.contentWindow) {
				return;
			}

			removeEventListener('message', handler);
			resolve();
		});
		document.body.appendChild(iframe);
		t.add_cleanup(() => iframe.remove());
	});

	return iframe.contentWindow;
};

const createPopup = async (t, win) => {
	const child = win.open(`${URL_DOMAIN}tmp/resources/test-frame.html`);

	await new Promise((resolve,reject) => {
		child.addEventListener('load', resolve);
		child.addEventListener('error', reject);
		t.add_cleanup(() => child.close());
	});

	return child;
};

async function send(win, message) {
	return new Promise((resolve) => {
		addEventListener('message', (event) => {
			resolve(event.data);
		});
		win.postMessage(message, '*');
	});
}

const dataUrl = () => `data:text/html,<meta charset="utf-8">
	<script src="${URL_DOMAIN}tmp/resources/utils.sub.js"><${''}/script>
	<script src="${URL_DOMAIN}tmp/resources/test-frame.js"><${''}/script>
	${Math.random()}`;

const objectUrl = (t) => {
	const encoder = new TextEncoder();
	const content = `<meta charset="utf-8">
	<script src="${URL_DOMAIN}tmp/resources/utils.sub.js"><${''}/script>
	<script src="${URL_DOMAIN}tmp/resources/test-frame.js"><${''}/script>
	${Math.random()}`;
	const url = URL.createObjectURL(
		new Blob([encoder.encode(content)], {type: 'text/html'})
	);

	t.add_cleanup(() => URL.revokeObjectURL(url));

	return url;
};

async function navigate(win, url) {
	return new Promise((resolve) => {
		addEventListener('message', function handler(event) {
			removeEventListener('message', handler);
			resolve();
		});
		win.location.assign(url);
	});
}

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	createMeta(parent, 'referrer', 'origin');
	parent.document.body.appendChild(child);

	assert_equals(await readReferer(child.contentWindow), URL_DOMAIN);
}, 'src-less iframe: inheritance');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	parent.document.body.appendChild(child);
	createMeta(parent, 'referrer', 'origin');

	assert_equals(await readReferer(child.contentWindow), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'src-less iframe: child independence from parent');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	parent.document.body.appendChild(child);
	createMeta(child.contentWindow, 'referrer', 'origin');

	assert_equals(await readReferer(child.contentWindow), URL_DOMAIN);
	assert_equals(await readReferer(parent), URL_COMPLETE);
}, 'src-less iframe: parent independence from child (child escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = document.createElement('iframe');

	createMeta(parent, 'referrer', 'origin');
	parent.document.body.appendChild(child);
	createMeta(child.contentWindow, 'referrer', 'unsafe-url');

	assert_equals(await readReferer(child.contentWindow), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'src-less iframe: parent independence from child (child de-escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t);
	createMeta(parent, 'referrer', 'origin');

	const child = await createPopup(t, parent);

	assert_equals(await readReferer(child), URL_DOMAIN);
}, 'popup: inheritance');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = await createPopup(t, parent);

	createMeta(parent, 'referrer', 'origin');

	assert_equals(await readReferer(child), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'popup: child independence from parent');

promise_test(async (t) => {
	const parent = await createIframe(t);
	const child = await createPopup(t, parent);

	createMeta(child, 'referrer', 'origin');

	assert_equals(await readReferer(child), URL_DOMAIN);
	assert_equals(await readReferer(parent), URL_COMPLETE);
}, 'popup: parent independence from child (child escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t);
	createMeta(parent, 'referrer', 'origin');
	const child = await createPopup(t, parent);

	createMeta(child, 'referrer', 'unsafe-url');

	assert_equals(await readReferer(child), URL_COMPLETE);
	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'popup: parent independence from child (child de-escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t, 'origin');

	await navigate(parent, dataUrl());

	assert_equals(await send(parent, 'readReferer'), URL_DOMAIN);
}, 'data: URL: inheritance');

promise_test(async (t) => {
	const parent = await createIframe(t);

	await navigate(parent, dataUrl());
	await send(parent, ['createMeta', 'referrer', 'origin']);
	await send(parent, 'back');

	assert_equals(await readReferer(parent), URL_COMPLETE);
}, 'data: URL: parent independence from child (child escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t, 'origin');

	await navigate(parent, dataUrl());
	await send(parent, ['createMeta', 'referrer', 'unsafe-url']);
	await send(parent, 'back');

	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'data: URL: parent independence from child (child de-escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t, 'origin');

	await navigate(parent, objectUrl(t));

	assert_equals(await send(parent, 'readReferer'), URL_DOMAIN);
}, 'blob: URL: inheritance');

promise_test(async (t) => {
	const parent = await createIframe(t);

	await navigate(parent, objectUrl(t));
	await send(parent, ['createMeta', 'referrer', 'origin']);
	await send(parent, 'back');

	assert_equals(await readReferer(parent), URL_COMPLETE);
}, 'blob: URL: parent independence from child (child escalates)');

promise_test(async (t) => {
	const parent = await createIframe(t, 'origin');

	await navigate(parent, objectUrl(t));
	await send(parent, ['createMeta', 'referrer', 'unsafe-url']);
	await send(parent, 'back');

	assert_equals(await readReferer(parent), URL_DOMAIN);
}, 'blob: URL: parent independence from child (child de-escalates)');
</script>
	</body>
</html>
